[FIX] base_user_role: fix field name ambiguity
Core Odoo defines a Roles field on res.partner, which is inherited on res.users.

This gives a recurring warning
```
Two fields (l10n_es_edi_facturae_ac_role_type_ids, role_ids) of res.users() have the same label: Roles. [Modules: None and base_user_role]
```

as well as real ambiguity in the list of views in search and export widgets.

https://github.com/odoo/odoo/blob/18.0/addons/l10n_es_edi_facturae/models/res_partner.py#L20-L21

Propose to change the field name to fix the ambiguity, and update labeling in other places
for consistency.